requirements.txt
appdirs==1.4.0
Django==1.10
django-autoslug==1.9.3
django-crispy-forms==1.6.1
django-eventtools==0.9.10
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==2.6.0
six==1.10.0
